[APG-666] Fetch full name from nomis user roles api
[APG-666] Remove manageUsersApiWebClient

[APG-666] Update fetch user detail response
