+++ helm_deploy/hmpps-accredited-programmes-api/values.yaml
@@ -47,8 +47,6 @@ generic-service:
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_MANAGE-OFFENCES-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_CASE-NOTES-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_CASE-NOTES-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
-      SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_MANAGE-USERS-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
-      SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_MANAGE-USERS-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_ALLOCATION-MANAGER-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_ALLOCATION-MANAGER-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
       SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_NOMIS-USER-ROLES-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
