+++ helm_deploy/values-preprod.yaml
@@ -16,7 +16,6 @@ generic-service:
     SERVICES_OASYS-API_BASE-URL: https://accredited-programmes-and-oasys-preprod.hmpps.service.justice.gov.uk
     SERVICES_MANAGE-OFFENCES-API_BASE-URL: https://manage-offences-api-preprod.hmpps.service.justice.gov.uk
     SERVICES_CASE-NOTES-API_BASE-URL: https://preprod.offender-case-notes.service.justice.gov.uk
-    SERVICES_MANAGE-USERS-API_BASE-URL: https://manage-users-api-preprod.hmpps.service.justice.gov.uk
     SERVICES_ALLOCATION-MANAGER-API_BASE-URL: https://preprod.moic.service.justice.gov.uk
     SERVICES_NOMIS-USER-ROLES-API_BASE-URL: https://nomis-user-roles-api-preprod.prison.service.justice.gov.uk
     SERVICES_PRISONER-ALERTS-API_BASE-URL: https://alerts-api-preprod.hmpps.service.justice.gov.uk
