+++ helm_deploy/values-prod.yaml
@@ -15,7 +15,6 @@ generic-service:
     SERVICES_OASYS-API_BASE-URL: https://accredited-programmes-and-oasys.hmpps.service.justice.gov.uk
     SERVICES_MANAGE-OFFENCES-API_BASE-URL: https://manage-offences-api.hmpps.service.justice.gov.uk
     SERVICES_CASE-NOTES-API_BASE-URL: https://offender-case-notes.service.justice.gov.uk
-    SERVICES_MANAGE-USERS-API_BASE-URL: https://manage-users-api.hmpps.service.justice.gov.uk
     SERVICES_ALLOCATION-MANAGER-API_BASE-URL: https://moic.service.justice.gov.uk
     SERVICES_NOMIS-USER-ROLES-API_BASE-URL: https://nomis-user-roles-api.prison.service.justice.gov.uk
     SERVICES_PRISONER-ALERTS-API_BASE-URL: https://alerts-api.hmpps.service.justice.gov.uk
