--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/manageusersApi/model/UserDetail.kt
@@ -1,10 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.client.manageusersApi.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-class UserDetail(
-  val username: String,
-  val name: String,
-  val active: Boolean,
-)
