+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/nomisUserRoleManagementApi/model/UserDetail.kt
@@ -9,4 +9,6 @@ data class UserDetail(
   val firstName: String,
   val lastName: String,
   val primaryEmail: String,
-)
+) {
+  fun fullName() = "$firstName $lastName"
+}
