+++ src/main/resources/application.yml
@@ -59,11 +59,6 @@ spring:
             client-id: hmpps-accredited-programmes-client-1
             client-secret: clientsecret
             authorization-grant-type: client_credentials
-          manage-users-api:
-            provider: hmpps-auth
-            client-id: hmpps-accredited-programmes-client-1
-            client-secret: clientsecret
-            authorization-grant-type: client_credentials
           allocation-manager-api:
             provider: hmpps-auth
             client-id: hmpps-accredited-programmes-client-1
@@ -143,8 +138,6 @@ services:
     base-url: http://localhost:9098
   case-notes-api:
     base-url: http://localhost:9099
-  manage-users-api:
-    base-url: http://localhost:9100
   allocation-manager-api:
     base-url: http://localhost:9101
   nomis-user-roles-api:
