+++ src/test/resources/application-test-h2.yml
@@ -59,8 +59,6 @@ services:
     base-url: http://localhost:#WIREMOCK_PORT
   case-notes-api:
     base-url: http://localhost:#WIREMOCK_PORT
-  manage-users-api:
-    base-url: http://localhost:#WIREMOCK_PORT
   allocation-manager-api:
     base-url: http://localhost:#WIREMOCK_PORT
   nomis-user-roles-api:
