+++ src/test/resources/application-test.yml
@@ -64,8 +64,6 @@ services:
     base-url: http://localhost:#WIREMOCK_PORT
   case-notes-api:
     base-url: http://localhost:#WIREMOCK_PORT
-  manage-users-api:
-    base-url: http://localhost:#WIREMOCK_PORT
   allocation-manager-api:
     base-url: http://localhost:#WIREMOCK_PORT
   nomis-user-roles-api:
