+++ src/test/resources/simulations/__files/userDetail.json
@@ -1,10 +1,8 @@
 {
   "username": "TEST_REFERRER_USER_1",
-  "active": false,
-  "name": "John Smith",
-  "authSource": "nomis",
-  "staffId": 231232,
-  "activeCaseLoadId": "MDI",
-  "userId": "231232",
-  "uuid": "5105a589-75b3-4ca0-9433-b96228c1c8f3"
+  "firstName": "John",
+  "lastName": "Smith",
+  "primaryEmail": "john.smith@test.com",
+  "staffId": 231232
 }
+
